requirements.txt
aiohttp
aiodns
aiosqlite
cchardet
pillow
pycups
toml

